# backend/openapi.json
{
  "components": {
    "responses": {
      "Unauthorized": {
        "description": "Authentication credentials are missing, malformed, expired, or invalid for this operation."
      }
    },
    "schemas": {
      "AlbumIndexState": {
        "enum": [
          "idle",
          "running",
          "completed",
          "canceled",
          "failed"
        ],
        "type": "string"
      },
      "AlbumIndexStatus": {
        "properties": {
          "cancelRequested": {
            "type": "boolean"
          },
          "failed": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "finishedAt": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "matched": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "processed": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "root": {
            "type": [
              "string",
              "null"
            ]
          },
          "scanned": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "startedAt": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "state": {
            "$ref": "#/components/schemas/AlbumIndexState"
          }
        },
        "required": [
          "state",
          "scanned",
          "matched",
          "processed",
          "failed",
          "cancelRequested"
        ],
        "type": "object"
      },
...
      "DeleteShare": {
        "properties": {
          "albumId": {
            "type": "string"
          },
          "shareId": {
            "type": "string"
          }
        },
        "required": [
          "albumId",
          "shareId"
        ],
        "type": "object"
      },
...
      "EditShare": {
        "properties": {
          "albumId": {
            "type": "string"
          },
          "share": {
            "$ref": "#/components/schemas/Share"
          }
        },
        "required": [
          "albumId",
          "share"
        ],
        "type": "object"
      },
...
      "IndexAlbumRequest": {
        "properties": {
          "album": {
            "type": "string"
          }
        },
        "required": [
          "album"
        ],
        "type": "object"
      },
      "IndexImageRequest": {
        "properties": {
          "album": {
            "type": [
              "string",
              "null"
            ]
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "image"
        ],
        "type": "object"
      },
...
      "RotateImageRequest": {
        "properties": {
          "asset_id": {
            "description": "Asset ID of the image to rotate",
            "type": "string"
          }
        },
        "required": [
          "asset_id"
        ],
        "type": "object"
      },
...
      "Share": {
        "properties": {
          "description": {
            "type": "string"
          },
          "exp": {
            "format": "int64",
            "type": "integer"
          },
          "password": {
            "type": [
              "string",
              "null"
            ]
          },
          "showDownload": {
            "type": "boolean"
          },
          "showMetadata": {
            "type": "boolean"
          },
          "showUpload": {
            "type": "boolean"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "description",
          "showMetadata",
          "showDownload",
          "showUpload",
          "exp"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "picasu",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/get/index/status": {
      "get": {
        "operationId": "get_album_index_status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumIndexStatus"
                }
              }
            },
            "description": "Album index status"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Report the state of the running album-indexing job.",
        "tags": [
          "index"
        ]
      }
    },
...
    "/post/index/album": {
      "post": {
        "description": "`album` is a path relative to `IMAGE_HOME` — use `\"/\"` for the root. Status\ncan be polled via `GET /get/index/status`.",
        "operationId": "index_album_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexAlbumRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Album indexing started"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Walk a directory under `IMAGE_HOME` and index all media files in the background.",
        "tags": [
          "index"
        ]
      }
    },
    "/post/index/cancel": {
      "post": {
        "operationId": "cancel_album_index_handler",
        "responses": {
          "200": {
            "description": "Album index cancelled"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Cancel a running album index job.",
        "tags": [
          "index"
        ]
      }
    },
    "/post/index/image": {
      "post": {
        "description": "Runs in the background; returns `202 Accepted` immediately.",
        "operationId": "index_image_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Image indexing started"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Index a single image by its path relative to `IMAGE_HOME`.",
        "tags": [
          "index"
        ]
      }
    },
...
    "/put/delete_share": {
      "put": {
        "operationId": "delete_share",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteShare"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Share deleted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Delete one share of an album.",
        "tags": [
          "albums"
        ]
      }
    },
...
    "/put/edit_share": {
      "put": {
        "operationId": "edit_share",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditShare"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Share updated"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Replace the share of an album.",
        "tags": [
          "albums"
        ]
      }
    },
...
    "/put/rotate-image": {
      "put": {
        "operationId": "rotate_image",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Image rotated"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Rotate an image and regenerate what depends on its orientation.",
        "tags": [
          "assets"
        ]
      }
    },
...
  }
}